Refactor security roles and update server-gateway image
Security roles for 'ADMIN', 'USER', and 'REPOSITORY' in the Spring Security configuration have been restructured for better organization and readability. The method names and comments have been updated, making the source code easier to understand and maintain. The server-gateway image was also updated in the deployment service YAML file to match the new jar file's location.